refactor: Reapply .cursorrules configuration and reduce 'use client' usage
